Make exp_any return the earliest result
Previously,
exp_any would return the match with the lowest index in the list.
So this:
exp_any(&vec![ReadUntil::String("hello"), ReadUntil::String("hell")]

With the input from the process as "hello",
will sometimes return "hell", and sometimes "hello".
Depending on how much the readbuffer happened to have collected.

With this change, the earliest match in the input will be matched.
That way, any difference in how much the read buffer has currently
consumed does not change the output of the exp_any function.

Signed-off-by: Erik Karlsson <triggger@gmail.com>