remove unneeded unwraps
Note, this slightly changes the pub API, so it should require a minor
version bump at least.

There is still an `unwrap` in a `Drop` where it's unclear if it
can happen or it can be handled better (maybe just ignored?) and
an `expect` in a function where the expect should not happen (it's
checked before), and I didn't consider that it's worth adding a new
Error kind for this in particular.

Signed-off-by: Petre Eftime <petre.eftime@gmail.com>