nix 0.30 (#145)
`dup`/`dup2` now encode their ownership semantics in the type system.
Special variants of `dup2` were added for the I/O redirection which we
use here. `dup` now takes an `Fd` instead of a `RawFd` and returns an
`OwnedFd`, which is directly convertible to a `File`, so we can get rid
of some conversion boilerplate in `get_file_handle`.